+++ lib/tinyagent/thread_item.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Tinyagent
-  # Abstract item on thread
   class ThreadItem < Model
     plugin :class_table_inheritance, key: :type
+    many_to_one :thread
   end
 end