+++ lib/tinyagent/tool_call.rb
@@ -1,8 +1,13 @@
 # frozen_string_literal: true
 
 module Tinyagent
-  # Save MCP configuration details.
   class ToolCall < Model
     many_to_one :message
+
+    def parsed_arguments
+      return nil unless arguments
+
+      JSON.parse(arguments)
+    end
   end
 end