+++ spec/support/factories.rb
@@ -1,8 +1,4 @@
 # frozen_string_literal: true
 
-autoload :BrainFactory, File.expand_path('factories/brain_factory', __dir__)
-autoload :DatabaseFactory, File.expand_path('factories/database_factory', __dir__)
-autoload :UserFactory, File.expand_path('factories/user_factory', __dir__)
-autoload :MessageFactory, File.expand_path('factories/message_factory', __dir__)
-autoload :RequestFactory, File.expand_path('factories/request_factory', __dir__)
-autoload :ChatFactory, File.expand_path('factories/chat_factory', __dir__)
+module Factories
+end